integrates fuzzing of buffered file with the fuzzing framework
